requirements.txt
openai
pinecone-client
google-auth
google-cloud-translate
